Refactored code, seperating related code into each of their related folders.
