+++ convex/_generated/api.d.ts
@@ -4,7 +4,6 @@
  *
  * THIS CODE IS AUTOMATICALLY GENERATED.
  *
- * Generated by convex@1.14.4.
  * To regenerate, run `npx convex dev`.
  * @module
  */
