+++ src/app/page.tsx
@@ -1,7 +1,6 @@
-
 export default function Home() {
   return (
-    <div>
+    <div className="flex-1 flex items-center justify-center text-3xl">
       Home
     </div>
   );
