+++ src/app/sign-up/[[...sign-up]]/page.tsx
@@ -2,7 +2,7 @@ import { SignUp } from '@clerk/nextjs'
 
 export default function Page() {
   return(
-    <div className="h-screen flex items-center justify-center">
+    <div className="flex-1 flex items-center justify-center">
       <SignUp />
     </div>
   )
