+++ src/components/hooks/useToggle.tsx
@@ -1,0 +1,6 @@
+import { useState } from "react";
+
+export function useToggle(initialValue = false) {
+  const [ value, setValue ] = useState(initialValue);
+  return [ value, setValue(!value) ]
+}
